Deprecate create_app() function
