+++ optuna_dashboard/__init__.py
@@ -1,5 +1,5 @@
-from .app import run_server  # noqa
-from .app import wsgi  # noqa
+from ._app import run_server  # noqa
+from ._app import wsgi  # noqa
 
 
 __version__ = "0.5.0"
