+++ optuna_dashboard/_cli.py
@@ -9,7 +9,7 @@ from optuna.storages import RDBStorage
 from optuna.storages import RedisStorage
 
 from . import __version__
-from .app import create_app
+from ._app import create_app
 
 
 AUTO_RELOAD = os.environ.get("OPTUNA_DASHBOARD_AUTO_RELOAD") == "1"
