+++ python_tests/test_serializers.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from optuna_dashboard.serializer import serialize_attrs
+from optuna_dashboard._serializer import serialize_attrs
 
 
 class SerializeAttrsTestCase(TestCase):
