+++ setup.cfg
@@ -35,7 +35,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    optuna-dashboard = optuna_dashboard.cli:main
+    optuna-dashboard = optuna_dashboard._cli:main
 
 [options.package_data]
 optuna_dashboard = public/*
